requirements.txt
langchain
langchain-groq
sentence-transformers
chromadb
PyMuPDF
streamlit
python-dotenv
tqdm
langchain-community
torch